requirements.txt
tensorflow>=2.12.0
keras>=2.12.0
scikit-learn>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.65.0
kaggle>=1.5.0
tensorflow-model-optimization>=0.7.0
plotly>=5.15.0
